_use fake configuration file as null object
