+++ protostar/composition_root.py
@@ -25,6 +25,7 @@ from protostar.commands.init.project_creator import (
 from protostar.compiler import ProjectCairoPathBuilder, ProjectCompiler
 from protostar.compiler.compiled_contract_reader import CompiledContractReader
 from protostar.configuration_file import ConfigurationFileFactory
+from protostar.io import InputRequester, log_color_provider
 from protostar.migrator import Migrator, MigratorExecutionEnvironment
 from protostar.protostar_cli import ProtostarCLI
 from protostar.protostar_toml import (
@@ -34,20 +35,13 @@ from protostar.protostar_toml import (
     ProtostarTOMLWriter,
     search_upwards_protostar_toml_path,
 )
+from protostar.self.protostar_directory import ProtostarDirectory, VersionManager
 from protostar.starknet_gateway import GatewayFacadeFactory
 from protostar.upgrader import (
     LatestVersionCacheTOML,
     LatestVersionChecker,
     LatestVersionRemoteChecker,
     UpgradeManager,
-)
-from protostar.io import (
-    InputRequester,
-    log_color_provider,
-)
-from protostar.self.protostar_directory import (
-    ProtostarDirectory,
-    VersionManager,
 )
