+++ protostar/configuration_file/configuration_file_factory.py
@@ -1,6 +1,7 @@
 from pathlib import Path
 from typing import Optional
 
+from protostar.configuration_file.fake_configuration_file import FakeConfigurationFile
 from protostar.protostar_exception import ProtostarException
 
 from .configuration_file import ConfigurationFile
@@ -15,10 +16,10 @@ class ConfigurationFileFactory:
         self._cwd = cwd
         self._active_profile_name = active_profile_name
 
-    def create(self) -> Optional[ConfigurationFile]:
+    def create(self) -> ConfigurationFile:
         protostar_toml_path = self._search_upwards_protostar_toml_path()
         if protostar_toml_path is None:
-            return None
+            return FakeConfigurationFile()
         protostar_toml_content = protostar_toml_path.read_text()
 
         configuration_file_v2 = self._create_configuration_toml_v2(
